Move mapConfig from BaseMapViewModel to AbstractMapContainerFragment
