+++ ground/src/main/java/com/google/android/ground/ui/common/ViewModelModule.kt
@@ -19,7 +19,6 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.google.android.ground.MainViewModel
 import com.google.android.ground.ui.datacollection.tasks.date.DateTaskViewModel
-import com.google.android.ground.ui.datacollection.tasks.location.CaptureLocationTaskMapViewModel
 import com.google.android.ground.ui.datacollection.tasks.location.CaptureLocationTaskViewModel
 import com.google.android.ground.ui.datacollection.tasks.multiplechoice.MultipleChoiceTaskViewModel
 import com.google.android.ground.ui.datacollection.tasks.number.NumberTaskViewModel
@@ -152,13 +151,6 @@ abstract class ViewModelModule {
   @IntoMap
   @ViewModelKey(MapTypeViewModel::class)
   abstract fun bindMapTypeViewModel(viewModel: MapTypeViewModel): ViewModel
-
-  @Binds
-  @IntoMap
-  @ViewModelKey(CaptureLocationTaskMapViewModel::class)
-  abstract fun bindCaptureLocationTaskMapViewModel(
-    viewModel: CaptureLocationTaskMapViewModel
-  ): ViewModel
 
   @Binds
   @IntoMap
