--- ground/src/test/java/com/google/android/ground/ui/datacollection/tasks/location/CaptureLocationTaskMapViewModelTest.kt
@@ -1,38 +1,0 @@
-/*
- * Copyright 2024 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.google.android.ground.ui.datacollection.tasks.location
-
-import com.google.android.ground.BaseHiltTest
-import com.google.android.ground.ui.common.MapConfig
-import com.google.common.truth.Truth.assertThat
-import dagger.hilt.android.testing.HiltAndroidTest
-import javax.inject.Inject
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.robolectric.RobolectricTestRunner
-
-@HiltAndroidTest
-@RunWith(RobolectricTestRunner::class)
-class CaptureLocationTaskMapViewModelTest : BaseHiltTest() {
-
-  @Inject lateinit var captureLocationTaskMapViewModel: CaptureLocationTaskMapViewModel
-
-  @Test
-  fun testGetMapConfig() {
-    assertThat(captureLocationTaskMapViewModel.mapConfig)
-      .isEqualTo(MapConfig(showOfflineImagery = true, allowGestures = false))
-  }
-}
